refactor: including useRef in linix use cases
