+++ src/pages/privileges/outlets/linixUseCase/adding-linix-use-case/forms/GeneralInformationForm/index.tsx
@@ -1,11 +1,11 @@
-import { useState } from "react";
-import { useFormik } from "formik";
+import { useState, forwardRef, useImperativeHandle } from "react";
+import { useFormik, FormikProps } from "formik";
 
 import { IMessageState } from "@pages/privileges/outlets/users/types/forms.types";
 import {
   IGeneralInformation,
   IHandleChangeFormData,
-} from "@pages/privileges/outlets/linixUseCase/adding-linix-use-case/index";
+} from "@pages/privileges/outlets/linixUseCase/adding-linix-use-case/types";
 
 import { GeneralInformationFormUI } from "./interface";
 import { generalMessage } from "../../config/messages.config";
@@ -22,112 +22,116 @@ interface GeneralInformationFormProps {
   readOnly?: boolean;
 }
 
-function GeneralInformationForm(props: GeneralInformationFormProps) {
-  const {
-    initialValues = {
-      n_Usecase: "",
-      n_Descrip: "",
-      i_Tipusec: "",
-      k_Funcio: "",
-      k_Opcion: "",
-    },
-    withSubmitButtons,
-    handleSubmit,
-    onHasChanges,
-    readOnly,
-    csOptions,
-    webOptions,
-  } = props;
+export const GeneralInformationForm = forwardRef(
+  function GeneralInformationForm(
+    props: GeneralInformationFormProps,
+    ref: React.Ref<FormikProps<IGeneralInformation>>
+  ) {
+    const {
+      initialValues = {
+        n_Usecase: "",
+        n_Descrip: "",
+        i_Tipusec: "",
+        k_Funcio: "",
+        k_Opcion: "",
+      },
+      withSubmitButtons,
+      handleSubmit,
+      onHasChanges,
+      readOnly,
+      csOptions,
+      webOptions,
+    } = props;
 
-  const [loading, setLoading] = useState(false);
-  const [message, setMessage] = useState<IMessageState>({
-    visible: false,
-  });
-
-  function onSubmit() {
-    setLoading(true);
-    setTimeout(() => {
-      handleSubmit(formik.values);
-      setLoading(false);
-      setMessage({
-        visible: true,
-        data: generalMessage.success,
-      });
-    }, LOADING_TIMEOUT);
-  }
-
-  const formik = useFormik({
-    initialValues,
-    validateOnChange: false,
-    onReset: () => {
-      if (onHasChanges) onHasChanges(false);
-    },
-    onSubmit,
-  });
-  const handleReset = () => {
-    if (onHasChanges) onHasChanges(false);
-  };
-
-  const handleSubmitForm = () => {
-    formik.validateForm().then((errors) => {
-      if (Object.keys(errors).length > 0) {
-        setMessage({
-          visible: true,
-          data: generalMessage.failed,
-        });
-      }
-      formik.handleSubmit();
-    });
-  };
-
-  const hasChanges = (valueCompare: IGeneralInformation) =>
-    JSON.stringify(initialValues) !== JSON.stringify(valueCompare);
-
-  const handleCloseSectionMessage = () => {
-    setMessage({
+    const [loading, setLoading] = useState(false);
+    const [message, setMessage] = useState<IMessageState>({
       visible: false,
     });
-  };
 
-  const handleChangeForm = (name: string, value: string) => {
-    const formikValues = {
-      ...formik.values,
-      [name]: value,
+    function onSubmit() {
+      setLoading(true);
+      setTimeout(() => {
+        handleSubmit(formik.values);
+        setLoading(false);
+        setMessage({
+          visible: true,
+          data: generalMessage.success,
+        });
+      }, LOADING_TIMEOUT);
+    }
+
+    const formik = useFormik({
+      initialValues,
+      validateOnChange: false,
+      onReset: () => {
+        if (onHasChanges) onHasChanges(false);
+      },
+      onSubmit,
+    });
+    const handleReset = () => {
+      if (onHasChanges) onHasChanges(false);
     };
 
-    if (onHasChanges) onHasChanges(hasChanges(formikValues));
-    formik.setFieldValue(name, value).then(() => {
-      if (withSubmitButtons) return;
+    const handleSubmitForm = () => {
       formik.validateForm().then((errors) => {
-        if (!errors || Object.keys(errors).length === 0) {
-          handleSubmit(formikValues);
+        if (Object.keys(errors).length > 0) {
           setMessage({
             visible: true,
-            data: generalMessage.success,
+            data: generalMessage.failed,
           });
         }
+        formik.handleSubmit();
       });
-    });
-  };
+    };
 
-  return (
-    <GeneralInformationFormUI
-      loading={loading}
-      handleReset={handleReset}
-      formik={formik}
-      message={message}
-      withSubmitButtons={withSubmitButtons}
-      handleCloseSectionMessage={handleCloseSectionMessage}
-      hasChanges={hasChanges}
-      onCloseSectionMessage={handleCloseSectionMessage}
-      formInvalid={formik.isValidating || formik.isValid}
-      handleSubmitForm={handleSubmitForm}
-      handleChangeForm={handleChangeForm}
-      readOnly={readOnly}
-      csOptions={csOptions}
-      webOptions={webOptions}
-    />
-  );
-}
+    const hasChanges = (valueCompare: IGeneralInformation) =>
+      JSON.stringify(initialValues) !== JSON.stringify(valueCompare);
 
-export { GeneralInformationForm };
+    const handleCloseSectionMessage = () => {
+      setMessage({
+        visible: false,
+      });
+    };
+
+    const handleChangeForm = (name: string, value: string) => {
+      const formikValues = {
+        ...formik.values,
+        [name]: value,
+      };
+
+      if (onHasChanges) onHasChanges(hasChanges(formikValues));
+      formik.setFieldValue(name, value).then(() => {
+        if (withSubmitButtons) return;
+        formik.validateForm().then((errors) => {
+          if (!errors || Object.keys(errors).length === 0) {
+            handleSubmit(formikValues);
+            setMessage({
+              visible: true,
+              data: generalMessage.success,
+            });
+          }
+        });
+      });
+    };
+
+    useImperativeHandle(ref, () => formik);
+    return (
+      <GeneralInformationFormUI
+        loading={loading}
+        handleReset={handleReset}
+        formik={formik}
+        message={message}
+        withSubmitButtons={withSubmitButtons}
+        handleCloseSectionMessage={handleCloseSectionMessage}
+        hasChanges={hasChanges}
+        onCloseSectionMessage={handleCloseSectionMessage}
+        formInvalid={formik.isValidating || formik.isValid}
+        handleSubmitForm={handleSubmitForm}
+        handleChangeForm={handleChangeForm}
+        readOnly={readOnly}
+        csOptions={csOptions}
+        webOptions={webOptions}
+      />
+    );
+  }
+);
